base: experiment with handling global account data as a separate processor
